virt_file_size: Use float in all places
Before there was a mix of int and float. This mix was now converted to float in
all places which is: Docs, Tests, Settings, CLI.

+++ cobbler/settings/migrations/V3_4_0.py
@@ -60,7 +60,7 @@ schema = Schema(
         Optional("default_template_type"): str,
         Optional("default_virt_bridge"): str,
         Optional("default_virt_disk_driver"): str,
-        Optional("default_virt_file_size"): int,
+        Optional("default_virt_file_size"): float,
         Optional("default_virt_ram"): int,
         Optional("default_virt_type"): str,
         Optional("enable_ipxe"): bool,
